+++ GUI/UserCreation.py
@@ -1,22 +1,23 @@
 #!/usr/bin/env python3
-from GUI.Dialogs.ConfirmTextDialog import ConfirmTextDialog
+
 import json
-from requests.api import head
-from wx.core import YES
 import os
-from Common.decorator import api_tool_decorator
 import Common.Globals as Globals
 import Utility.EventUtility as eventUtil
+import csv
+import wx
+import wx.grid
+
+from Common.decorator import api_tool_decorator
+
+from GUI.Dialogs.ConfirmTextDialog import ConfirmTextDialog
+
+from Utility.API.UserUtility import createNewUser, deleteUser, modifyUser
 from Utility.Resource import (
     createNewFile,
     displayMessageBox,
     postEventToFrame,
 )
-import csv
-import wx
-import wx.grid
-
-from Utility.API.UserUtility import createNewUser, deleteUser, modifyUser
 
 
 class UserCreation(wx.Frame):
