+++ Utility/API/AppUtilities.py
@@ -3,7 +3,6 @@
 import esperclient
 import time
 import wx
-
 import Common.Globals as Globals
 import Utility.EventUtility as eventUtil
 
@@ -15,7 +14,6 @@ from Utility.Logging.ApiToolLogging import ApiToolLog
 from Utility.API.CommandUtility import (
     executeCommandOnDevice,
     executeCommandOnGroup,
-    postEsperCommand,
 )
 from Utility.Resource import (
     displayMessageBox,
@@ -350,62 +348,6 @@ def getAllAppVersionsForHost(
         raise Exception(
             "Exception when calling ApplicationApi->get_app_versions: %s\n" % e
         )
-
-
-@api_tool_decorator()
-def clearAppData(frame, device):
-    json_resp = None
-    try:
-        appToUse = frame.sidePanel.selectedAppEntry["pkgName"]
-        _, apps = getdeviceapps(
-            device.id, createAppList=False, useEnterprise=Globals.USE_ENTERPRISE_APP
-        )
-        cmdArgs = {}
-        for app in apps["results"]:
-            if app["package_name"] == appToUse:
-                cmdArgs["package_name"] = app["package_name"]
-                cmdArgs["application_name"] = app["app_name"]
-                cmdArgs["version_code"] = app["version_code"]
-                cmdArgs["version_name"] = app["version_name"]
-                if app["app_type"] == "GOOGLE":
-                    cmdArgs["is_g_play"] = True
-                else:
-                    cmdArgs["is_g_play"] = False
-                break
-
-        if cmdArgs:
-            reqData = {
-                "command_type": "DEVICE",
-                "command_args": cmdArgs,
-                "devices": [device.id],
-                "groups": [],
-                "device_type": Globals.CMD_DEVICE_TYPE,
-                "command": "CLEAR_APP_DATA",
-            }
-            resp, json_resp = postEsperCommand(reqData)
-            logBadResponse(resp.request.url, resp, json_resp)
-            if resp.status_code > 300:
-                postEventToFrame(eventUtil.myEVT_ON_FAILED, device)
-            if resp.status_code < 300:
-                frame.Logging(
-                    "---> Clear %s App Data Command has been sent to %s"
-                    % (cmdArgs["application_name"], device.device_name)
-                )
-        else:
-            frame.Logging(
-                "ERROR: Failed to send Clear %s App Data Command to %s"
-                % (
-                    frame.sidePanel.selectedAppEntry["name"],
-                    device.device_name,
-                )
-            )
-    except Exception as e:
-        ApiToolLog().LogError(e)
-        frame.Logging(
-            "ERROR: Failed to send Clear App Data Command to %s" % (device.device_name)
-        )
-        postEventToFrame(eventUtil.myEVT_ON_FAILED, device)
-    return json_resp
 
 
 def getApplication(application_id):
